+++ src/Commands/Base/TokenHandler.cs
@@ -1,5 +1,4 @@
-﻿using AngleSharp.Text;
-using Microsoft.Identity.Client;
+﻿using Microsoft.Identity.Client;
 using Microsoft.SharePoint.Client;
 using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Model;
@@ -8,6 +7,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Management.Automation;
 using System.Net.Http;
+using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 
@@ -46,33 +46,58 @@ namespace PnP.PowerShell.Commands.Base
         /// Returns the permission scopes of the oAuth JWT token being passed in
         /// </summary>
         /// <param name="accessToken">The oAuth JWT token</param>
-        /// <returns>String array containing the scopes</returns>
+        /// <returns>String array containing the scopes in the format <audience>/<scope></returns>
         internal static string[] ReturnScopes(string accessToken)
         {
             var decodedToken = new JwtSecurityToken(accessToken);
 
+            // Retrieve the audience the token was issued for
+            var audience = decodedToken.Audiences.FirstOrDefault();
+
             // The scopes can either be stored in the roles or scp claim, so we examine both
-            var scopes = decodedToken.Claims.Where(c => c.Type == "roles" || c.Type == "scp").SelectMany(r => r.Value.Split(" "));
+            var scopes = decodedToken.Claims.Where(c => c.Type == "roles" || c.Type == "scp").SelectMany(r => r.Value.Split(" ").Select(r => $"{audience}/{r}"));
             return scopes.ToArray();
         }
 
         /// <summary>
-        /// Ensures the oAuth JWT token holds the permissions in it (roles) that match with with the required permissions for the cmdlet provided through the attributes on the cmdlet
+        /// Ensures the oAuth JWT token holds the permissions in it (roles) that match with with the required permissions for the cmdlet provided through the attributes on the cmdlet that have the same audience
         /// </summary>
         /// <param name="cmdletType">The cmdlet that will be executed. Used to check for the permissions attribute.</param>
         /// <param name="accessToken">The oAuth JWT token that needs to be validated for its roles</param>
-        /// <param name="throwExceptionIfPermissionsNotPresent">Have this function throw an exception if the required permissions in the permission attribute do not match with the available permissions in the access token (true). Defaults to false.</param>
         /// <exception cref="PSArgumentException">Thrown if the permissions set through the permissions attribute do not match the roles in the JWT token</exception>
-        internal static void EnsureRequiredPermissionsAvailableInAccessToken(Type cmdletType, string accessToken)
+        internal static void EnsureRequiredPermissionsAvailableInAccessTokenAudience(Cmdlet cmdlet, string accessToken)
         {
-            var permissionEvaluationResponses = AccessTokenPermissionValidationResponse.EvaluatePermissions(cmdletType, accessToken);
-            foreach (var permissionEvaluationResponse in permissionEvaluationResponses)
+            // Decode the JWT token
+            var decodedToken = new JwtSecurityToken(accessToken);
+
+            // Retrieve the audience the token was issued for
+            var audience = decodedToken.Audiences.FirstOrDefault();            
+
+            // Validate the permissions in the access token against the permissions required for the cmdlet through the attributes provided at the class level of the cmdlet
+            var permissionEvaluationResponses = AccessTokenPermissionValidationResponse.EvaluatePermissions(cmdlet, accessToken, audience);
+
+            // If any of the permission evaluations has the required permissions present, we can return and are good to go permission-wise
+            if (permissionEvaluationResponses.Any(p => p.RequiredPermissionsPresent)) return;
+
+            // Retrieve the scopes we have in our AccessToken
+            var scopes = ReturnScopes(accessToken);
+
+            // None of the permission attributes matched the permissions in the access token, so we throw an exception
+            var exceptionTextBuilder = new StringBuilder();
+            exceptionTextBuilder.AppendLine($"Authorization Denied: token misses the following permission scope(s) on the {audience} audience:");
+
+            for (int i = 0; i < permissionEvaluationResponses.Length; i++)
             {
-                if (permissionEvaluationResponse.RequiredPermissionsPresent) return;
+                exceptionTextBuilder.AppendLine($"{string.Join(" and ", permissionEvaluationResponses[i].MissingPermissions.Select(s => s[(audience != null ? audience.Length + 1 : 0)..]))}");
+
+                if(i < permissionEvaluationResponses.Length - 1)
+                {
+                    exceptionTextBuilder.AppendLine(" or ");
+                }
             }
-            var missingPermissions = permissionEvaluationResponses.Select(r => r.MissingPermissions);
-            throw new PSArgumentException($"Authorization Denied: Token used does not contain permission scope{(missingPermissions.Count() != 1 ? "s" : "")} for one or more of these permissions: {string.Concat(missingPermissions, ", ")}");
-        }
+
+            throw new PSArgumentException(exceptionTextBuilder.ToString());
+        }        
 
         /// <summary>
         /// Returns an oAuth JWT access token
@@ -107,71 +132,6 @@ namespace PnP.PowerShell.Commands.Base
                         }
 
                         accessToken = authManager.GetAccessToken(audience);
-                        var currentScopes = ReturnScopes(accessToken);
-
-                        var scopesPresent = false;
-
-                        if (cmdlet != null && cmdlet.GetType() != null)
-                        {
-                            var requiredListedScopes = (RequiredMinimalApiPermissions[])Attribute.GetCustomAttributes(cmdlet.GetType(), typeof(RequiredMinimalApiPermissions));
-                            // assume these are 'or'
-                            if (requiredListedScopes != null)
-                            {
-                                foreach (var requiredListedScope in requiredListedScopes)
-                                {
-                                    var matchedScopes = 0;
-                                    // check if scopes are present
-                                    foreach (var permission in requiredListedScope.PermissionScopes)
-                                    {
-                                        if (currentScopes.Contains(permission, StringComparison.InvariantCultureIgnoreCase))
-                                        {
-                                            matchedScopes++;
-                                        }
-                                    }
-                                    if (matchedScopes == requiredListedScope.PermissionScopes.Length)
-                                    {
-                                        scopesPresent = true;
-                                        var requiredScopes = requiredListedScope.PermissionScopes;
-                                        if (contextSettings.Type != Framework.Utilities.Context.ClientContextType.AzureADCertificate)
-                                        {
-                                            accessToken = authManager.GetAccessTokenAsync(requiredScopes).GetAwaiter().GetResult();
-                                            // we have a match, jump out of the loop
-                                            break;
-                                        }
-                                    }
-                                }
-
-                                if (!scopesPresent)
-                                {
-                                    var requiredScopes = requiredListedScopes.Length > 0 ? requiredListedScopes.First().PermissionScopes : null;
-
-                                    if (contextSettings.Type == Framework.Utilities.Context.ClientContextType.AzureADCertificate)
-                                    {
-                                        requiredScopes = new[] { audience }; // override for app only
-                                    }
-                                    if (requiredScopes == null && !string.IsNullOrEmpty(audience))
-                                    {
-                                        requiredScopes = new[] { audience };
-                                    }
-
-                                    cmdlet.WriteVerbose($"Acquiring oAuth token for {(requiredScopes.Length != 1 ? requiredScopes.Length + " " : "")}permission scope{(requiredScopes.Length != 1 ? "s" : "")} {string.Join(", ", requiredScopes)}");
-                                    accessToken = authManager.GetAccessTokenAsync(requiredScopes).GetAwaiter().GetResult();
-
-                                    // Retrieve the scopes from the access token
-                                    var scopes = ReturnScopes(accessToken);
-
-                                    // Perform logging on the scopes in the retrieved access token
-                                    if (scopes.Length > 0)
-                                    {
-                                        cmdlet.WriteVerbose($"Access token acquired containing the following {(scopes.Length != 1 ? scopes.Length + " " : "")}{RetrieveTokenType(accessToken).ToString().ToLowerInvariant()} scope{(scopes.Length == 1 ? "" : "s")}: {string.Join(", ", scopes)}");
-                                    }
-                                    else
-                                    {
-                                        cmdlet.WriteVerbose($"No scopes could be determined from the access token");
-                                    }
-                                }
-                            }
-                        }
                     }
                 }
             }
@@ -179,21 +139,6 @@ namespace PnP.PowerShell.Commands.Base
             {
                 cmdlet.WriteVerbose($"Unable to acquire token for resource {connection.GraphEndPoint}");
                 return null;
-            }
-
-            var permissionEvaluations = AccessTokenPermissionValidationResponse.EvaluatePermissions(cmdlet.GetType(), accessToken);
-            var permissionsInPlace = false;
-            foreach (var permissionEvaluation in permissionEvaluations)
-            {
-                if (permissionEvaluation.MissingPermissions.Length == 0)
-                {
-                    permissionsInPlace = true;
-                    break;
-                }
-            }
-            if (!permissionsInPlace)
-            {
-                cmdlet.WriteVerbose($"The currect access token might not have the required permissions to execute this cmdlet. Required are one or more of the following: {string.Join(", ", permissionEvaluations.Select(p => p.MissingPermissions))}");
             }
 
             return accessToken;